Fix performance issues in styling
