+++ packages/graph-explorer/src/modules/EdgesStyling/index.ts
@@ -1,0 +1,2 @@
+export { default as EdgesStyling } from "./EdgesStyling";
+export * from "./EdgeStyleDialog";
