+++ packages/graph-explorer/src/modules/GraphViewer/index.ts
@@ -1,2 +1,1 @@
 export { default } from "./GraphViewer";
-export type { GraphViewerProps } from "./GraphViewer";
