+++ packages/graph-explorer/src/modules/NodesStyling/NodesStyling.tsx
@@ -11,16 +11,10 @@ import useTranslations from "@/hooks/useTranslations";
 import SingleNodeStyling from "./SingleNodeStyling";
 import { Fragment } from "react/jsx-runtime";
 import { SidebarCloseButton } from "../SidebarCloseButton";
+import { Virtuoso } from "react-virtuoso";
+import NodeStyleDialog from "./NodeStyleDialog";
 
-export type NodesStylingProps = {
-  onNodeCustomize(nodeType?: string): void;
-  customizeNodeType?: string;
-};
-
-function NodesStyling({
-  customizeNodeType,
-  onNodeCustomize,
-}: NodesStylingProps) {
+function NodesStyling() {
   const vtConfigs = useDisplayVertexTypeConfigs().values().toArray();
   const t = useTranslations();
 
@@ -33,21 +27,21 @@ function NodesStyling({
         </PanelHeaderActions>
       </PanelHeader>
       <PanelContent className="gap-2">
-        {vtConfigs.map((vtConfig, index) => {
-          return (
+        <Virtuoso
+          className="h-full grow"
+          data={vtConfigs}
+          itemContent={(index, vtConfig) => (
             <Fragment key={vtConfig.type}>
               {index !== 0 ? <Divider /> : null}
 
               <SingleNodeStyling
                 vertexType={vtConfig.type}
-                opened={customizeNodeType === vtConfig.type}
-                onOpen={() => onNodeCustomize(vtConfig.type)}
-                onClose={() => onNodeCustomize(undefined)}
                 className="px-3 pb-3 pt-2"
               />
             </Fragment>
-          );
-        })}
+          )}
+        />
+        <NodeStyleDialog />
       </PanelContent>
     </Panel>
   );
