+++ packages/graph-explorer/src/modules/NodesStyling/index.ts
@@ -1,0 +1,2 @@
+export { default as NodesStyling } from "./NodesStyling";
+export * from "./NodeStyleDialog";
